# server/swagger-output.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "0.1.0",
    "title": "오픈마켓 API",
    "description": "오픈마켓 API Server입니다.<br>\n      <a href=\"/\">버전별 변경사항 확인</a>"
  },
  "servers": [
    {
      "url": "https://api.fesp.shop",
      "description": ""
    },
    {
      "url": "http://localhost",
      "description": "로컬 테스트"
    }
  ],
...
  "paths": {
...
    "/products/": {
      "get": {
        "tags": [
          "상품"
        ],
        "summary": "상품 목록 조회",
        "description": "상품 목록을 조회합니다.<br>지원되는 검색 조건 이외의 속성으로 검색할 경우 custom 파라미터를 이용하면 됩니다.<br>custom 파라미터에 지정하는 값은 MongoDB 검색어로 사용되며, api 클라이언트 프로그램의 샘플 예제 이외에 더 자세한 검색 조건 지정은 다음 문서를 참고하세요.<br><a target=\"_blank\" href=\"https://www.mongodb.com/docs/drivers/node/current/fundamentals/crud/query-document\">검색어 지정 방법</a>",
        "parameters": [
          {
            "name": "minPrice",
            "description": "최저 가격",
            "in": "query",
            "default": 0,
            "example": 10000,
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "maxPrice",
            "description": "최고 가격",
            "in": "query",
            "default": 99999999999,
            "example": 50000,
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "minShippingFees",
            "description": "최저 배송비",
            "in": "query",
            "default": 0,
            "example": 2500,
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "maxShippingFees",
            "description": "최고 배송비",
            "in": "query",
            "default": 99999999999,
            "example": 5000,
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "keyword",
            "description": "상품명 검색어",
            "in": "query",
            "example": "레고",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "seller_id",
            "description": "판매자 id",
            "in": "query",
            "example": 2,
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "custom",
            "description": "custom 검색 조건",
            "in": "query",
            "example": "{'extra.isNew': true}",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "description": "페이지",
            "in": "query",
            "example": 2,
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "limit",
            "description": "한 페이지당 항목 수<br>   지정하지 않으면 최대 100개 까지 응답    ",
            "in": "query",
            "example": 10,
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "sort",
            "description": "정렬(내림차순: -1, 오름차순: 1)",
            "in": "query",
            "example": "{'price': -1}",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "showSoldOut",
            "description": "매진 상품 포함 여부",
            "in": "query",
            "example": true,
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "성공",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/productListRes"
                }
              }
            }
          },
          "500": {
            "description": "서버 에러",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error500"
                }
              }
            }
          }
        }
      }
    },
    "/products/{_id}": {
      "get": {
        "tags": [
          "상품"
        ],
        "summary": "상품 상세 조회",
        "description": "상품 상세 정보를 조회합니다.",
        "parameters": [
          {
            "name": "_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "상품 id",
            "example": 4
          }
        ],
        "responses": {
          "200": {
            "description": "성공",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/productInfoRes"
                }
              }
            }
          },
          "404": {
            "description": "리소스가 존재하지 않음",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error404"
                }
              }
            }
          },
          "500": {
            "description": "서버 에러",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error500"
                }
              }
            }
          }
        }
      }
    },
...
    "/orders/{_id}/products/{product_id}": {
      "patch": {
        "tags": [
          "구매"
        ],
        "summary": "상품별 주문 상태 수정",
        "description": "상품별로 주문 상태를 수정합니다.(반품 요청, 환불 요청 등)<br>여러 판매자의 상품을 한번에 주문하고 결제했을 경우 하나의 주문에 여러 상품이 있고 각 상품별로 주문 상태를 관리하고 싶을 때 사용합니다.",
        "parameters": [
          {
            "name": "_id",
            "description": "구매 id",
            "in": "path",
            "example": 2,
            "required": true,
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "product_id",
            "description": "상품 id",
            "in": "path",
            "example": 3,
            "required": true,
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "성공",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/updateOrderProductRes"
                }
              }
            }
          },
          "401": {
            "description": "인증 실패",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error401"
                }
              }
            }
          },
          "404": {
            "description": "본인의 구매 id가 아니거나 존재하지 않는 구매 id",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error404"
                }
              }
            }
          },
          "500": {
            "description": "서버 에러",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error500"
                }
              }
            }
          }
        },
        "security": [
          {
            "Access Token": []
          }
        ],
        "requestBody": {
          "description": "주문 상태 정보",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#components/schemas/updateOrderProduct"
              }
            }
          }
        }
      }
    },
...
    "/replies/products/{_id}": {
      "get": {
        "tags": [
          "구매 후기"
        ],
        "summary": "상품 구매 후기 목록",
        "description": "지정한 상품의 구매 후기 목록을 조회합니다.",
        "parameters": [
          {
            "name": "_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "상품 id",
            "example": 3
          },
          {
            "name": "rating",
            "description": "후기 점수",
            "in": "query",
            "example": 5,
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "sort",
            "description": "정렬(내림차순: -1, 오름차순: 1)",
            "in": "query",
            "example": "{'replies': -1}",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "성공",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/replyListRes"
                }
              }
            }
          },
          "500": {
            "description": "서버 에러",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error500"
                }
              }
            }
          }
        }
      }
    },
...
    "/replies/seller/{seller_id}": {
      "get": {
        "tags": [
          "구매 후기"
        ],
        "summary": "판매자 구매 후기 목록",
        "description": "판매자의 상품별로 등록된 모든 구매 후기 목록을 조회합니다.",
        "parameters": [
          {
            "name": "seller_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "판매자 id",
            "example": 2
          }
        ],
        "responses": {
          "200": {
            "description": "성공",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/sellerReplyListRes"
                }
              }
            }
          },
          "500": {
            "description": "서버 에러",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error500"
                }
              }
            }
          }
        }
      }
    },
...
    "/posts/seller/{_id}": {
      "get": {
        "tags": [
          "게시판"
        ],
        "summary": "판매자의 상품에 등록된 게시글 목록",
        "description": "판매자의 상품에 등록된 게시글 목록을 조회합니다.",
        "parameters": [
          {
            "name": "_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "판매자 id",
            "example": 2
          },
          {
            "name": "product_id",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "product_id",
            "description": "상품 id",
            "in": "path",
            "example": 1,
            "required": true,
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "type",
            "description": "게시판 종류",
            "in": "query",
            "default": "post",
            "example": "qna",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "keyword",
            "description": "검색어<br>제목과 내용 검색에 사용되는 키워드",
            "in": "query",
            "example": "배송",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "custom",
            "description": "custom 검색 조건",
            "in": "query",
            "example": "{'createdAt': {'$gte': '2024.04', '$lt': '2024.05'}}",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "description": "페이지",
            "in": "query",
            "example": 2,
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "limit",
            "description": "한 페이지당 항목 수",
            "in": "query",
            "example": 10,
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "sort",
            "description": "정렬(내림차순: -1, 오름차순: 1)",
            "in": "query",
            "example": "{'createdAt': 1}",
            "default": "{'createdAt': -1}",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "성공",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/postListRes"
                }
              }
            }
          },
          "500": {
            "description": "서버 에러",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error500"
                }
              }
            }
          }
        }
      }
    },
...
    "/seller/products/": {
      "get": {
        "tags": [
          "상품 관리"
        ],
        "summary": "판매 상품 목록 조회",
        "description": "자신의 판매 상품 목록을 조회합니다.",
        "parameters": [
          {
            "name": "minPrice",
            "description": "최저 가격",
            "in": "query",
            "default": 0,
            "example": 10000,
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "maxPrice",
            "description": "최고 가격",
            "in": "query",
            "default": 99999999999,
            "example": 50000,
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "minShippingFees",
            "description": "최저 배송비",
            "in": "query",
            "default": 0,
            "example": 0,
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "maxShippingFees",
            "description": "최고 배송비",
            "in": "query",
            "default": 99999999999,
            "example": 5000,
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "keyword",
            "description": "상품명 검색어",
            "in": "query",
            "example": "레고",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "seller_id",
            "description": "판매자 id",
            "in": "query",
            "example": 2,
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "custom",
            "description": "custom 검색 조건",
            "in": "query",
            "example": "{'extra.isBest': true}",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "description": "페이지",
            "in": "query",
            "example": 2,
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "limit",
            "description": "한 페이지당 항목 수",
            "in": "query",
            "example": 10,
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "sort",
            "description": "정렬(내림차순: -1, 오름차순: 1)",
            "in": "query",
            "example": "{'price': -1}",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "성공",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/myProductListRes"
                }
              }
            }
          },
          "401": {
            "description": "인증 실패",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error401"
                }
              }
            }
          },
          "500": {
            "description": "서버 에러",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error500"
                }
              }
            }
          }
        },
        "security": [
          {
            "Access Token": []
          }
        ]
      },
      "post": {
        "tags": [
          "상품 관리"
        ],
        "summary": "상품 등록",
        "description": "상품을 등록합니다.<br>상품 등록 후 상품 정보를 반환합니다.",
        "responses": {
          "201": {
            "description": "성공",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/productCreateRes"
                }
              }
            }
          },
          "401": {
            "description": "인증 실패",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error401"
                }
              }
            }
          },
          "422": {
            "description": "입력값 검증 오류",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error422"
                }
              }
            }
          },
          "500": {
            "description": "서버 에러",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error500"
                }
              }
            }
          }
        },
        "security": [
          {
            "Access Token": []
          }
        ],
        "requestBody": {
          "description": "상품 정보가 저장된 객체입니다.<br>price: 가격(필수)<br>quantity: 수량(필수)<br>name: 상품명(필수)<br>content: 상품 설명(필수)<br>shippingFees: 배송비(선택, 생략시 0원)<br>mainImages: 상품 설명 이미지(선택)<br>show: 상품 노출 여부(선택, 기본값 true). false로 지정될 경우 구매 회원의 상품 목록 조회에 노출되지 않음.<br>extra: 추가 데이터(선택). 추가하고 싶은 아무 속성이나 지정",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#components/schemas/productCreate"
              }
            }
          }
        }
      }
    },
...
    "/seller/orders/{_id}/products/{product_id}": {
      "patch": {
        "tags": [
          "주문 관리"
        ],
        "summary": "상품별 주문 상태 수정",
        "description": "상품별로 주문 상태를 수정합니다.(배송 시작, 환불 완료 등)<br>여러 판매자의 상품을 한번에 주문하고 결제했을 경우 하나의 주문에 여러 상품이 있고 각 상품별로 주문 상태를 관리하고 싶을 때 사용합니다.",
        "parameters": [
          {
            "name": "_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "주문 id",
            "example": 2
          },
          {
            "name": "product_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "상품 id",
            "example": 3
          }
        ],
        "responses": {
          "200": {
            "description": "성공",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/updateOrderProductSellerRes"
                }
              }
            }
          },
          "401": {
            "description": "인증 실패",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error401"
                }
              }
            }
          },
          "404": {
            "description": "리소스가 존재하지 않음",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error404"
                }
              }
            }
          },
          "500": {
            "description": "서버 에러",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error500"
                }
              }
            }
          }
        },
        "security": [
          {
            "Access Token": []
          }
        ],
        "requestBody": {
          "description": "수정할 주문 정보가 저장된 객체입니다.<br>state: 주문한 상품의 state 값으로 지정됩니다.<br>나머지 속성은 주문한 상품의 history 속성에 추가됩니다.",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/updateOrderProductSellerBody"
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "Access Token": {
        "type": "http",
        "scheme": "bearer",
        "in": "header",
        "bearerFormat": "JWT"
      },
      "Refresh Token": {
        "type": "http",
        "scheme": "bearer",
        "in": "header",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "error404": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "number",
            "example": 0
          },
          "message": {
            "type": "string",
            "example": "/api/xxx 리소스를 찾을 수 없습니다."
          }
        }
      },
...
      "error500": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "number",
            "example": 0
          },
          "message": {
            "type": "string",
            "example": "요청하신 작업 처리에 실패했습니다. 잠시 후 다시 이용해 주시기 바랍니다."
          }
        }
      },
...
      "updateOrderProductRes": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "number",
            "example": 1
          },
          "item": {
            "type": "object",
            "properties": {
              "_id": {
                "type": "number",
                "example": 2
              },
              "product_id": {
                "type": "number",
                "example": 3
              },
              "state": {
                "type": "string",
                "example": "OS110"
              },
              "memo": {
                "type": "string",
                "example": "레고 클래식 상품이 동작하지 않습니다. 반품 요청합니다."
              },
              "updatedAt": {
                "type": "string",
                "example": "2024.04.08 07:55:34"
              }
            }
          }
        }
      },
...
    },
...
  }
}